refactor: split proxy settings helpers
