+++ gpt2giga/core/config/_settings/__init__.py
@@ -1,0 +1,1 @@
+"""Internal building blocks for ``gpt2giga.core.config.settings``."""
